% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_sane_parameterication}
\alias{assert_sane_parameterication}
\title{Sane Parameter Check}
\usage{
assert_sane_parameterication(P, a, b)
}
\arguments{
\item{P}{Modulus used to devide n, iff n is equivalent to (0 mod P).}

\item{a}{Factor by which to multiply n.}

\item{b}{Value to add to the scaled value of n.}
}
\description{
Handles the sanity check for the parameterisation (P,a,b)
}
\details{
Required by both the function and reverse function, to assert that they
have sane parameters, otherwise will force stop the execution.
}
