% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseFunction.R
\name{reverseFunction}
\alias{reverseFunction}
\title{The "inverse"/"reverse" Collatz function}
\usage{
reverseFunction(n, P = 2, a = 3, b = 1)
}
\arguments{
\item{n}{(numeric|bigz|bigq) The value on which
to perform the reverse Collatz function}

\item{P}{(numeric|bigz|bigq) Modulus used to divide
n, iff n is equivalent to (0 mod P) Default is 2.}

\item{a}{(numeric|bigz|bigq) Factor by which to multiply n. Default is 3.}

\item{b}{(numeric|bigz|bigq) Value to add
to the scaled value of n. Default is 1.}
}
\value{
A vector of either numeric or bigz | bigq type
}
\description{
Calculates the values that would return the input under the Collatz function.
}
\details{
Returns the output of a single application of a Collatz-esque reverse
function. If only one value is returned, it is the value that would be
divided by P. If two values are returned, the first is the value that
would be divided by P, and the second value is that which would undergo
the multiply and add step, regardless of which is larger.
}
