% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stoppingTimeTerminus}
\alias{stoppingTimeTerminus}
\title{Stopping Time Terminus}
\usage{
stoppingTimeTerminus(n, total_stop)
}
\arguments{
\item{n}{The initial value to confirm against a stopping time check.}

\item{total_stop}{If false, the lambda will confirm that iterations of n
have reached the oriented stopping time to reach a value closer to 0.
If true, the lambda will simply check equality to 1.}
}
\value{
An anonymous function to check for the stopping time.
}
\description{
Provides the appropriate lambda to use to check if iterations on an initial
value have reached either the stopping time, or total stopping time.
}
