% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collatz.R
\docType{package}
\name{collatz}
\alias{collatz}
\title{Collatz}
\description{
Functions related to the Collatz/Syracuse/3N+1 problem.
}
\details{
Provides the basic functionality to interact with the Collatz conjecture.
The parameterisation uses the same (P,a,b) notation as Conway's generalisations.
Besides the function and reverse function, there is also functionality to retrieve
the hailstone sequence, the "stopping time"/"total stopping time", or tree-graph.
The only restriction placed on parameters is that both P and a can't be 0.
}
